# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Summer School",
        "version": "1.0"
    },
    "paths": {
        "/api/v1/session/create": {
            "post": {
                "tags": [
                    "Session"
                ],
                "description": "Create payment session",
                "operationId": "App\\Http\\Swagger\\PaymentAPIController::createSessionId",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "sum",
                                    "details"
                                ],
                                "properties": {
                                    "sum": {
                                        "description": "Сумма платежа",
                                        "type": "numeric"
                                    },
                                    "details": {
                                        "description": "Назначение платежа",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create payment session",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "success": "true",
                                        "data": {
                                            "sessionId": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/pay": {
            "post": {
                "tags": [
                    "Transaction"
                ],
                "description": "Create transaction",
                "operationId": "App\\Http\\Swagger\\PaymentAPIController::pay",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "card_number",
                                    "cvc",
                                    "dates",
                                    "sessionId"
                                ],
                                "properties": {
                                    "card_number": {
                                        "description": "Номер карты (4111 1111 1111 1234)",
                                        "type": "string"
                                    },
                                    "cvc": {
                                        "description": "CVV/CVC (111)",
                                        "type": "string"
                                    },
                                    "dates": {
                                        "description": "Срок действия карты (01/11)",
                                        "type": "string"
                                    },
                                    "sessionId": {
                                        "description": "Session ID",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create payment session",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "success": "true",
                                        "data": "Транзакция прошла успешно!"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/transactions": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "description": "Get transaction",
                "operationId": "App\\Http\\Swagger\\PaymentAPIController::getTransactions",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "description": "От (08/07/2020)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "До (09/07/2020)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Obtain transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "success": "true",
                                        "data": [
                                            {
                                                "id": 1,
                                                "card_number": "1234",
                                                "cvc": "1234",
                                                "dates": "01/24",
                                                "status": "1",
                                                "session_id": 10,
                                                "created_at": "2020-07-08 13:34:53",
                                                "updated_at": "2020-07-08 13:34:53"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/user/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "description": "Login",
                "operationId": "App\\Http\\Swagger\\UserAPIController::login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create payment session",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "success": "true",
                                        "data": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}